Fix issues from PR review
