+++ analytics/models.py
@@ -1,4 +1,4 @@
-rom django.db import models
+from django.db import models
 from course.models.models import Submission, Question, UserQuestionJunction
 from course.fields import JSONField
