+++ analytics/models/parsons.py
@@ -9,7 +9,7 @@ class ParsonsSubmissionAnalytics(SubmissionAnalytics):
     blank_lines = models.IntegerField(default=0)
     comment_lines = models.IntegerField(default=0)
     import_lines = models.IntegerField(default=0)
-    cc = models.IntegerField(default=0)
+    cyclomatic_complexity = models.IntegerField(default=0)
     method = models.IntegerField(default=0)
     operator = models.IntegerField(default=0)
     operand = models.IntegerField(default=0)
