+++ analytics/services/submission_analytics.py
@@ -70,7 +70,7 @@ def get_all_submission_analytics():
                                                            blank_lines=sub_analytics_dict.blank_lines,
                                                            comment_lines=sub_analytics_dict.comment_lines,
                                                            import_lines=sub_analytics_dict.imported_lines,
-                                                           cc=sub_analytics_dict.cc,
+                                                           cyclomatic_complexity=sub_analytics_dict.cc,
                                                            method=sub_analytics_dict.method,
                                                            operator=sub_analytics_dict.operator,
                                                            operand=sub_analytics_dict.operand,
@@ -101,7 +101,7 @@ def get_all_submission_analytics():
                                                               blank_lines=sub_analytics_dict.blank_lines,
                                                               comment_lines=sub_analytics_dict.comment_lines,
                                                               import_lines=sub_analytics_dict.imported_lines,
-                                                              cc=sub_analytics_dict.cc,
+                                                              cyclomatic_complexity=sub_analytics_dict.cc,
                                                               method=sub_analytics_dict.method,
                                                               operator=sub_analytics_dict.operator,
                                                               operand=sub_analytics_dict.operand,
@@ -171,7 +171,7 @@ def create_submission_analytics(submission):
                                                                           blank_lines=sub_analytics_dict.blank_lines,
                                                                           comment_lines=sub_analytics_dict.comment_lines,
                                                                           import_lines=sub_analytics_dict.imported_lines,
-                                                                          cc=sub_analytics_dict.cc,
+                                                                          cyclomatic_complexity=sub_analytics_dict.cc,
                                                                           method=sub_analytics_dict.method,
                                                                           operator=sub_analytics_dict.operator,
                                                                           operand=sub_analytics_dict.operand,
@@ -201,7 +201,7 @@ def create_submission_analytics(submission):
                                                                              blank_lines=sub_analytics_dict.blank_lines,
                                                                              comment_lines=sub_analytics_dict.comment_lines,
                                                                              import_lines=sub_analytics_dict.imported_lines,
-                                                                             cc=sub_analytics_dict.cc,
+                                                                             cyclomatic_complexity=sub_analytics_dict.cc,
                                                                              method=sub_analytics_dict.method,
                                                                              operator=sub_analytics_dict.operator,
                                                                              operand=sub_analytics_dict.operand,
