+++ api/urls.py
@@ -44,7 +44,7 @@ router.register(r'admin', AdminViewSet, basename='admin')
 router.register(r'course-admin', CourseAdminViewSet, basename='admin-course')
 router.register(r'uqj-update', UpdateUQJViewSet, basename='uqj-update')
 router.register(r'question-report', QuestionReportViewSet, basename='question-report')
-router.register(r'submission-analytics', AnalyticsViewSet, basename='submission-analytics')
+router.register(r'analytics/submission', AnalyticsViewSet, basename='submission-analytics')
 
 app_name = 'api'
 urlpatterns = [
